test: restore isolated Voxtral contract test
